# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS API
  version: 1.0.0
  description: |
    A RESTful Learning Management System API built with Node.js, Express.js,
    MongoDB, and Mongoose.

    The API provides authentication, user management, course management,
    lessons, enrollments, progress tracking, reviews, favorites,
    instructor dashboard functionality, and Stripe payments.
servers:
  - url: http://localhost:4000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      description: Course ID
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f1a2b3c4d5e6f789012345
    LessonId:
      name: lessonId
      in: path
      required: true
      description: Lesson ID
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f1a2b3c4d5e6f789012346
...
    Page:
      name: page
      in: query
      required: false
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    Limit:
      name: limit
      in: query
      required: false
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        example: 10
  schemas:
...
    InstructorSummary:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        firstName:
          type: string
          example: Ahmed
        lastName:
          type: string
          example: Mohamed
        email:
          type: string
          format: email
          example: ahmed@example.com
...
    Course:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        title:
          type: string
          example: Node.js Basics
        description:
          type: string
          example: Learn the fundamentals of Node.js
        price:
          type: number
          minimum: 0
          example: 199
        category:
          type: string
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
        instructor:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        published:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CoursePopulated:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        title:
          type: string
          example: Node.js Basics
        description:
          type: string
          example: Learn the fundamentals of Node.js
        price:
          type: number
          minimum: 0
          example: 199
        category:
          type: string
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
        instructor:
          $ref: '#/components/schemas/InstructorSummary'
        published:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCourseRequest:
      type: object
      required:
        - title
        - description
        - price
        - category
      properties:
        title:
          type: string
          example: Node.js Basics
        description:
          type: string
          example: Learn the fundamentals of Node.js
        price:
          type: number
          minimum: 0
          example: 199
        category:
          type: string
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
          example: Advanced Node.js
        description:
          type: string
          example: Advanced Node.js backend concepts
        price:
          type: number
          minimum: 0
          example: 299
        category:
          type: string
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
        published:
          type: boolean
          example: true
    CourseFilters:
      type: object
      properties:
        category:
          type: string
          nullable: true
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
        search:
          type: string
          nullable: true
          example: node
        sort:
          type: string
          nullable: true
          enum:
            - price
            - title
            - createdAt
          example: createdAt
        order:
          type: string
          nullable: true
          enum:
            - asc
            - desc
          example: desc
    CourseLessonSummary:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012347
        title:
          type: string
          example: Introduction
        order:
          type: integer
          example: 1
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CoursePopulated'
        pagination:
          type: object
          properties:
            totalCourses:
              type: integer
              example: 20
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            totalPages:
              type: integer
              example: 2
        filters:
          $ref: '#/components/schemas/CourseFilters'
    MyCoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CoursePopulated'
    CourseDetailsResponse:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/CoursePopulated'
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/CourseLessonSummary'
    DeleteCourseResponse:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
        message:
          type: string
          example: Course deleted successfully
...
    Progress:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012351
        student:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        lesson:
          type: string
          example: 65f1a2b3c4d5e6f789012348
        course:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        completed:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProgressResponse:
      type: object
      properties:
        progress:
          $ref: '#/components/schemas/Progress'
    CourseProgressResponse:
      type: object
      properties:
        progressPercentage:
          type: number
          minimum: 0
          maximum: 100
          example: 75
...
    Favorite:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012353
        student:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        course:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FavoriteCourseSummary:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        title:
          type: string
          example: Node.js Basics
        category:
          type: string
          enum:
            - programming
            - design
            - marketing
            - business
            - languages
            - personal-development
          example: programming
    FavoritePopulated:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012353
        student:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        course:
          $ref: '#/components/schemas/FavoriteCourseSummary'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FavoriteResponse:
      type: object
      properties:
        favorite:
          $ref: '#/components/schemas/Favorite'
    FavoritesResponse:
      type: object
      properties:
        favorites:
          type: array
          items:
            $ref: '#/components/schemas/FavoritePopulated'
...
    CourseResponse:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
...
    DeleteFavoriteResponse:
      type: object
      properties:
        favorite:
          $ref: '#/components/schemas/Favorite'
    Error:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Something went wrong
paths:
...
  /api/courses/:
    get:
      tags:
        - Courses
      operationId: getAllCourses
      summary: Get published courses
      description: |
        Returns published courses with pagination.

        Supports category filtering, title search, sorting, and pagination.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: category
          in: query
          description: Filter courses by category
          schema:
            type: string
            enum:
              - programming
              - design
              - marketing
              - business
              - languages
              - personal-development
            example: programming
        - name: search
          in: query
          description: Search courses by title
          schema:
            type: string
            example: Node
        - name: sort
          in: query
          description: Field used for sorting
          schema:
            type: string
            enum:
              - price
              - title
              - createdAt
            example: createdAt
        - name: order
          in: query
          description: Sorting direction
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
            example: desc
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '400':
          description: Invalid sorting parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/courses/create:
    post:
      tags:
        - Courses
      operationId: createCourse
      summary: Create a new course
      description: Creates a new course for the authenticated instructor or administrator.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Invalid course data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/courses/myCourses:
    get:
      tags:
        - Courses
      operationId: getMyCourses
      summary: Get current user's courses
      description: Returns courses created by the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyCoursesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/courses/{courseId}:
    get:
      tags:
        - Courses
      operationId: getCourseById
      summary: Get course by ID
      description: |
        Returns course details.

        Published courses include lesson summaries.
        Unpublished courses can only be accessed by the course owner
        or an administrator.
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailsResponse'
        '403':
          description: This course is not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Courses
      operationId: updateCourse
      summary: Update course
      description: Course owner or administrator can update the course.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Courses
      operationId: deleteCourse
      summary: Delete course
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCourseResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/courses/{courseId}/lessons/{lessonId}/complete:
    post:
      tags:
        - Progress
      operationId: markLessonComplete
      summary: Mark a lesson as completed
      description: Marks the specified lesson as completed for the authenticated student.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/LessonId'
      responses:
        '201':
          description: Lesson marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
        '400':
          description: Lesson does not belong to the specified course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Lesson is already marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/courses/{courseId}/progress:
    get:
      tags:
        - Progress
      operationId: getCourseProgress
      summary: Get course progress
      description: Returns the authenticated student's progress percentage for the course.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course progress retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseProgressResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No lessons found for this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/favorites/:
    get:
      tags:
        - Favorites
      operationId: getMyFavorites
      summary: Get current user's favorite courses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Favorites retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/favorites/{courseId}:
    post:
      tags:
        - Favorites
      operationId: addFavorite
      summary: Add course to favorites
      description: Adds a published course to the authenticated student's favorites.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '201':
          description: Course added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied or course is not available for favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Course is already in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Favorites
      operationId: removeFavorite
      summary: Remove course from favorites
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course removed from favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteFavoriteResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...